Split dialog feature into its own files
The multi-character dialog API is conceptually distinct from single-character
streaming TTS, and the combined file was getting hard to navigate (~1700
lines). Split into three focused units:

- ElevenLabsHttp.h: header-only shared HTTP plumbing — ElevenLabsCall RAII
  wrapper, checkResponse<T>() helper, captureRequestIdHeader / appendToString
  callbacks, kNoBodyForLog sentinel. Lives in creatures::voice::elevenlabs_http.
- DialogClient.h / .cpp: new DialogClient class with generateDialog(),
  forcedAlignment(), static stripTags(). Owns the Dialog* and ForcedAlignment*
  types. Spans renamed to "DialogClient.*".
- StreamingTTSClient.h / .cpp: removed the dialog/forced-alignment surface,
  removed stripTags, removed the local helper anon namespace (now in
  ElevenLabsHttp.h). generateSpeechREST now uses the shared helpers via
  using-declarations.

The refactor benefit is preserved: the shared envelope (auth header,
request-id capture, curl rc + http-code mapping) is still in one place,
just at a higher level so both clients can use it.

CMakeLists.txt globs src/server/voice/* so the new files are picked up
automatically. Clean build under -Wshadow -Wall -Wextra -Wpedantic.
