+++ src/server/voice/StreamingTTSClient.cpp
@@ -3,7 +3,6 @@
 #include <arpa/inet.h>
 #include <cstring>
 #include <netdb.h>
-#include <optional>
 #include <random>
 #include <sstream>
 #include <sys/socket.h>
@@ -16,6 +15,7 @@
 #include <openssl/err.h>
 #include <openssl/ssl.h>
 
+#include "ElevenLabsHttp.h"
 #include "server/namespace-stuffs.h"
 #include "util/ObservabilityManager.h"
 
@@ -739,164 +739,11 @@ Result<StreamingTTSResult> StreamingTTSClient::generateSpeech(const std::string 
     return ttsResult;
 }
 
-// ---------------------------------------------------------------------------
-// Shared HTTP plumbing for the ElevenLabs REST endpoints (generateSpeechREST,
-// generateDialog, forcedAlignment). All three share the same envelope —
-// xi-api-key auth, request-id capture, curl rc + http-code → Result<T> mapping
-// — so it lives in this anon namespace; each method below still owns its own
-// content-type / body / response-parsing logic.
-// ---------------------------------------------------------------------------
 namespace {
-
-/// Header callback that captures the request-id (or x-request-id) value into
-/// the std::string pointed to by `userdata`. Case-insensitive match.
-size_t captureRequestIdHeader(char *data, size_t size, size_t nmemb, void *userdata) {
-    auto *requestId = static_cast<std::string *>(userdata);
-    const size_t totalBytes = size * nmemb;
-    std::string header(data, totalBytes);
-
-    std::string lower = header;
-    std::transform(lower.begin(), lower.end(), lower.begin(),
-                   [](unsigned char c) { return static_cast<char>(std::tolower(c)); });
-
-    if (lower.find("request-id:") == 0 || lower.find("x-request-id:") == 0) {
-        const auto colonPos = header.find(':');
-        if (colonPos != std::string::npos) {
-            std::string value = header.substr(colonPos + 1);
-            while (!value.empty() && (value.front() == ' ' || value.front() == '\t')) {
-                value.erase(value.begin());
-            }
-            while (!value.empty() && (value.back() == '\r' || value.back() == '\n' || value.back() == ' ')) {
-                value.pop_back();
-            }
-            *requestId = value;
-        }
-    }
-    return totalBytes;
-}
-
-/// Append-to-std::string write callback for one-shot JSON responses.
-size_t appendToString(char *data, size_t size, size_t nmemb, void *userdata) {
-    auto *buf = static_cast<std::string *>(userdata);
-    const size_t totalBytes = size * nmemb;
-    buf->append(data, totalBytes);
-    return totalBytes;
-}
-
-/// RAII wrapper for an ElevenLabs HTTP call.
-///
-/// Owns the curl easy handle, the headers list, and (optionally) the MIME body —
-/// all freed in the destructor regardless of how the call site exits. Pre-wires
-/// the standard xi-api-key auth header and the request-id capture callback so
-/// every call site gets those for free. The two TTS endpoints actually use the
-/// captured request-id; forced-alignment ignores it (the cost of always wiring
-/// the header callback is one trivial branch per response header line).
-///
-/// Call sites still own the request-specific bits — Content-Type, body
-/// (POSTFIELDS or MIMEPOST via createMime()), write callback, and timeout — so
-/// the file stays readable about what each request actually sends and parses.
-class ElevenLabsCall {
-  public:
-    ElevenLabsCall(const std::string &apiKey, const std::string &url) {
-        curl_ = curl_easy_init();
-        if (!curl_) {
-            return;
-        }
-        addHeader(fmt::format("xi-api-key: {}", apiKey));
-        curl_easy_setopt(curl_, CURLOPT_URL, url.c_str());
-        curl_easy_setopt(curl_, CURLOPT_HEADERFUNCTION, &captureRequestIdHeader);
-        curl_easy_setopt(curl_, CURLOPT_HEADERDATA, &requestId_);
-    }
-
-    ~ElevenLabsCall() {
-        if (mime_) {
-            curl_mime_free(mime_);
-        }
-        if (headers_) {
-            curl_slist_free_all(headers_);
-        }
-        if (curl_) {
-            curl_easy_cleanup(curl_);
-        }
-    }
-
-    ElevenLabsCall(const ElevenLabsCall &) = delete;
-    ElevenLabsCall &operator=(const ElevenLabsCall &) = delete;
-    ElevenLabsCall(ElevenLabsCall &&) = delete;
-    ElevenLabsCall &operator=(ElevenLabsCall &&) = delete;
-
-    [[nodiscard]] bool initOk() const { return curl_ != nullptr; }
-    [[nodiscard]] CURL *handle() const { return curl_; }
-    [[nodiscard]] const std::string &requestId() const { return requestId_; }
-
-    void addHeader(const std::string &header) { headers_ = curl_slist_append(headers_, header.c_str()); }
-
-    /// Create + own a curl_mime against this handle. Returns the handle so the
-    /// caller can add parts. Ownership stays with this wrapper (freed in dtor).
-    curl_mime *createMime() {
-        mime_ = curl_mime_init(curl_);
-        return mime_;
-    }
-
-    /// Execute. Applies any added headers and the MIME body if one was created,
-    /// then performs the request. On return, `httpCode` is set from the response;
-    /// the return value is the curl rc.
-    CURLcode perform(long &httpCode) {
-        if (headers_) {
-            curl_easy_setopt(curl_, CURLOPT_HTTPHEADER, headers_);
-        }
-        if (mime_) {
-            curl_easy_setopt(curl_, CURLOPT_MIMEPOST, mime_);
-        }
-        const CURLcode rc = curl_easy_perform(curl_);
-        httpCode = 0;
-        curl_easy_getinfo(curl_, CURLINFO_RESPONSE_CODE, &httpCode);
-        return rc;
-    }
-
-  private:
-    CURL *curl_ = nullptr;
-    curl_slist *headers_ = nullptr;
-    curl_mime *mime_ = nullptr;
-    std::string requestId_;
-};
-
-/// Map curl rc + http code into a Result<T> error, or std::nullopt on HTTP 200.
-///
-/// - curl rc != OK → InternalError (DNS, connect, timeout, TLS, ...)
-/// - HTTP 400 → InvalidData (the caller's request was malformed)
-/// - any other non-200 → InternalError (upstream failure)
-///
-/// `whatFailed` prefixes the error message ("ElevenLabs dialog"); `bodyForLog`
-/// is appended to non-200 errors so we capture the API's complaint in the log.
-template <typename T>
-std::optional<Result<T>> checkElevenLabsResponse(CURLcode rc, long httpCode, const std::string &whatFailed,
-                                                 const std::string &bodyForLog,
-                                                 const std::shared_ptr<OperationSpan> &span) {
-    if (rc != CURLE_OK) {
-        std::string msg = fmt::format("{} curl error: {}", whatFailed, curl_easy_strerror(rc));
-        error(msg);
-        if (span) {
-            span->setError(msg);
-        }
-        return Result<T>{ServerError(ServerError::InternalError, msg)};
-    }
-    if (httpCode != 200) {
-        std::string msg = fmt::format("{} HTTP {}: {}", whatFailed, httpCode, bodyForLog);
-        error(msg);
-        if (span) {
-            span->setError(msg);
-        }
-        const auto code = (httpCode == 400) ? ServerError::InvalidData : ServerError::InternalError;
-        return Result<T>{ServerError(code, msg)};
-    }
-    return std::nullopt;
-}
-
-/// Sentinel that means "no body to log" — used when the upstream body is the
-/// streaming NDJSON which we've already consumed and trimmed.
-constexpr const char *kNoBodyForLog = "";
-
+using elevenlabs_http::appendToString;
+using elevenlabs_http::checkResponse;
+using elevenlabs_http::ElevenLabsCall;
+using elevenlabs_http::kNoBodyForLog;
 } // namespace
 
 Result<StreamingTTSResult> StreamingTTSClient::generateSpeechREST(const std::string &apiKey, const std::string &voiceId,
@@ -1048,8 +895,7 @@ Result<StreamingTTSResult> StreamingTTSClient::generateSpeechREST(const std::str
     const CURLcode res = call.perform(httpCode);
     result.requestId = call.requestId();
 
-    if (auto err =
-            checkElevenLabsResponse<StreamingTTSResult>(res, httpCode, "ElevenLabs REST TTS", kNoBodyForLog, span)) {
+    if (auto err = checkResponse<StreamingTTSResult>(res, httpCode, "ElevenLabs REST TTS", kNoBodyForLog, span)) {
         return *err;
     }
 
@@ -1084,360 +930,6 @@ Result<StreamingTTSResult> StreamingTTSClient::generateSpeechREST(const std::str
     }
 
     return result;
-}
-
-// ---------------------------------------------------------------------------
-// Phase 1: multi-character dialog + forced-alignment primitives.
-// These are the building blocks for the new /api/v1/animation/dialog endpoint.
-// generateDialog is intentionally NOT subject to the eleven_v3 blocklist that
-// the ad-hoc single-character path enforces — text-to-dialogue is eleven_v3 only.
-// ---------------------------------------------------------------------------
-
-std::string StreamingTTSClient::stripTags(const std::string &text) {
-    // Drop "[...]" tags, then collapse runs of whitespace to single spaces, then trim.
-    std::string noTags;
-    noTags.reserve(text.size());
-    int depth = 0;
-    for (char c : text) {
-        if (c == '[') {
-            ++depth;
-            continue;
-        }
-        if (c == ']') {
-            if (depth > 0) {
-                --depth;
-            }
-            continue;
-        }
-        if (depth == 0) {
-            noTags.push_back(c);
-        }
-    }
-
-    std::string collapsed;
-    collapsed.reserve(noTags.size());
-    bool inSpace = false;
-    for (char c : noTags) {
-        const bool isSpace = (c == ' ' || c == '\t' || c == '\n' || c == '\r' || c == '\f' || c == '\v');
-        if (isSpace) {
-            if (!inSpace && !collapsed.empty()) {
-                collapsed.push_back(' ');
-            }
-            inSpace = true;
-        } else {
-            collapsed.push_back(c);
-            inSpace = false;
-        }
-    }
-    while (!collapsed.empty() && collapsed.back() == ' ') {
-        collapsed.pop_back();
-    }
-    return collapsed;
-}
-
-Result<DialogResult> StreamingTTSClient::generateDialog(const std::string &apiKey,
-                                                        const std::vector<DialogInput> &inputs,
-                                                        const std::string &outputFormat,
-                                                        std::shared_ptr<OperationSpan> parentSpan) {
-    auto span = creatures::observability->createChildOperationSpan("StreamingTTSClient.generateDialog", parentSpan);
-    if (span) {
-        span->setAttribute("dialog.inputs", static_cast<int64_t>(inputs.size()));
-        span->setAttribute("audio.format", outputFormat);
-        std::size_t totalChars = 0;
-        for (const auto &in : inputs) {
-            totalChars += in.text.size();
-        }
-        span->setAttribute("dialog.total_chars", static_cast<int64_t>(totalChars));
-    }
-
-    if (inputs.empty()) {
-        std::string msg = "generateDialog requires at least one input turn";
-        if (span)
-            span->setError(msg);
-        return Result<DialogResult>{ServerError(ServerError::InvalidData, msg)};
-    }
-
-    // ---- Build the request body. Text-to-dialogue is eleven_v3-only (other models
-    // are rejected by the server with HTTP 400 "does not support dialogue"). The
-    // ad-hoc single-character path BLOCKLISTS eleven_v3; we deliberately bypass
-    // that here — it's the whole point of this method.
-    nlohmann::json body;
-    body["model_id"] = "eleven_v3";
-    auto inputsArr = nlohmann::json::array();
-    for (const auto &in : inputs) {
-        inputsArr.push_back({{"voice_id", in.voiceId}, {"text", in.text}});
-    }
-    body["inputs"] = std::move(inputsArr);
-    const std::string bodyStr = body.dump();
-
-    const std::string url =
-        fmt::format("https://api.elevenlabs.io/v1/text-to-dialogue/with-timestamps?output_format={}", outputFormat);
-
-    // Single-blob JSON response (NOT newline-delimited), so we accumulate to a
-    // string and parse once at the end.
-    std::string respBuf;
-    DialogResult result;
-    result.audioFormat = outputFormat;
-
-    ElevenLabsCall call(apiKey, url);
-    if (!call.initOk()) {
-        std::string msg = "Failed to initialize curl";
-        if (span)
-            span->setError(msg);
-        return Result<DialogResult>{ServerError(ServerError::InternalError, msg)};
-    }
-    call.addHeader("Content-Type: application/json");
-    call.addHeader("Accept: application/json");
-    curl_easy_setopt(call.handle(), CURLOPT_POSTFIELDS, bodyStr.c_str());
-    curl_easy_setopt(call.handle(), CURLOPT_POSTFIELDSIZE, static_cast<long>(bodyStr.size()));
-    curl_easy_setopt(call.handle(), CURLOPT_WRITEFUNCTION, &appendToString);
-    curl_easy_setopt(call.handle(), CURLOPT_WRITEDATA, &respBuf);
-    // Dialog is slow — eleven_v3 with forced-alignment downstream is the bottleneck.
-    // 90s gives headroom for ~2000-char scenes without leaving the call open forever.
-    curl_easy_setopt(call.handle(), CURLOPT_TIMEOUT, 90L);
-
-    long httpCode = 0;
-    const CURLcode res = call.perform(httpCode);
-    result.requestId = call.requestId();
-
-    if (auto err = checkElevenLabsResponse<DialogResult>(res, httpCode, "ElevenLabs dialog", respBuf, span)) {
-        return *err;
-    }
-
-    // ---- Parse the response.
-    nlohmann::json json;
-    try {
-        json = nlohmann::json::parse(respBuf);
-    } catch (const nlohmann::json::exception &e) {
-        std::string msg = fmt::format("ElevenLabs dialog: response was not valid JSON: {}", e.what());
-        error(msg);
-        if (span)
-            span->setError(msg);
-        return Result<DialogResult>{ServerError(ServerError::InternalError, msg)};
-    }
-
-    // audio_base64 → bytes
-    if (json.contains("audio_base64") && json["audio_base64"].is_string()) {
-        try {
-            const auto decoded = base64::from_base64(json["audio_base64"].get<std::string>());
-            result.audioData.assign(reinterpret_cast<const uint8_t *>(decoded.data()),
-                                    reinterpret_cast<const uint8_t *>(decoded.data()) + decoded.size());
-        } catch (const std::exception &e) {
-            std::string msg = fmt::format("ElevenLabs dialog: base64 decode failed: {}", e.what());
-            error(msg);
-            if (span)
-                span->setError(msg);
-            return Result<DialogResult>{ServerError(ServerError::InternalError, msg)};
-        }
-    } else {
-        std::string msg = "ElevenLabs dialog: response missing audio_base64";
-        error(msg);
-        if (span)
-            span->setError(msg);
-        return Result<DialogResult>{ServerError(ServerError::InternalError, msg)};
-    }
-
-    // alignment.characters → kept for downstream sanity checks. The TIMES in
-    // alignment are broken on eleven_v3 (confirmed empirically) — we don't expose
-    // them on the result and we won't use them. Real timing comes from forcedAlignment().
-    if (json.contains("alignment") && json["alignment"].is_object()) {
-        const auto &al = json["alignment"];
-        if (al.contains("characters") && al["characters"].is_array()) {
-            try {
-                result.alignmentCharacters = al["characters"].get<std::vector<std::string>>();
-            } catch (const nlohmann::json::exception &) {
-                // Non-string entries — treat as fatal; the downstream index math relies on this.
-                std::string msg = "ElevenLabs dialog: alignment.characters was not an array of strings";
-                error(msg);
-                if (span)
-                    span->setError(msg);
-                return Result<DialogResult>{ServerError(ServerError::InternalError, msg)};
-            }
-        }
-    }
-
-    // voice_segments[] — char ranges + speaker. Times kept for diagnostics only.
-    if (json.contains("voice_segments") && json["voice_segments"].is_array()) {
-        for (const auto &seg : json["voice_segments"]) {
-            DialogVoiceSegment v;
-            if (seg.contains("voice_id"))
-                v.voiceId = seg["voice_id"].get<std::string>();
-            if (seg.contains("character_start_index"))
-                v.characterStartIndex = seg["character_start_index"].get<std::size_t>();
-            if (seg.contains("character_end_index"))
-                v.characterEndIndex = seg["character_end_index"].get<std::size_t>();
-            if (seg.contains("dialogue_input_index"))
-                v.dialogInputIndex = seg["dialogue_input_index"].get<std::size_t>();
-            if (seg.contains("start_time_seconds"))
-                v.startTimeSeconds = seg["start_time_seconds"].get<double>();
-            if (seg.contains("end_time_seconds"))
-                v.endTimeSeconds = seg["end_time_seconds"].get<double>();
-            result.voiceSegments.push_back(std::move(v));
-        }
-    }
-    if (result.voiceSegments.empty()) {
-        std::string msg = "ElevenLabs dialog: response missing voice_segments";
-        error(msg);
-        if (span)
-            span->setError(msg);
-        return Result<DialogResult>{ServerError(ServerError::InternalError, msg)};
-    }
-
-    // Duration estimate. pcm_48000 is mono 16-bit @ 48 kHz → 96000 bytes/sec.
-    // Generic pcm_* fallback assumes the rate in the format string.
-    if (outputFormat == "pcm_48000") {
-        result.audioDurationSeconds = static_cast<double>(result.audioData.size()) / 96000.0;
-    } else if (outputFormat == "pcm_44100") {
-        result.audioDurationSeconds = static_cast<double>(result.audioData.size()) / 88200.0;
-    } else if (outputFormat == "pcm_24000") {
-        result.audioDurationSeconds = static_cast<double>(result.audioData.size()) / 48000.0;
-    } else if (outputFormat.find("mp3") != std::string::npos) {
-        // Same heuristic the existing REST path uses; only approximate.
-        result.audioDurationSeconds = static_cast<double>(result.audioData.size()) / 24000.0;
-    }
-
-    info("Dialog complete: {} inputs, {} bytes audio, {} alignment chars, {} segments, "
-         "~{:.2f}s, request_id={}",
-         inputs.size(), result.audioData.size(), result.alignmentCharacters.size(), result.voiceSegments.size(),
-         result.audioDurationSeconds, result.requestId);
-
-    if (span) {
-        span->setAttribute("audio.bytes", static_cast<int64_t>(result.audioData.size()));
-        span->setAttribute("audio.duration_s", result.audioDurationSeconds);
-        span->setAttribute("alignment.chars", static_cast<int64_t>(result.alignmentCharacters.size()));
-        span->setAttribute("dialog.segments", static_cast<int64_t>(result.voiceSegments.size()));
-        span->setAttribute("request_id", result.requestId);
-        span->setSuccess();
-    }
-    return result;
-}
-
-Result<ForcedAlignmentResult> StreamingTTSClient::forcedAlignment(const std::string &apiKey,
-                                                                  const std::vector<uint8_t> &audio,
-                                                                  const std::string &contentType,
-                                                                  const std::string &transcript,
-                                                                  std::shared_ptr<OperationSpan> parentSpan) {
-    auto span = creatures::observability->createChildOperationSpan("StreamingTTSClient.forcedAlignment", parentSpan);
-    if (span) {
-        span->setAttribute("audio.bytes", static_cast<int64_t>(audio.size()));
-        span->setAttribute("audio.content_type", contentType);
-        span->setAttribute("transcript.length", static_cast<int64_t>(transcript.size()));
-    }
-
-    if (audio.empty()) {
-        std::string msg = "forcedAlignment requires non-empty audio";
-        if (span)
-            span->setError(msg);
-        return Result<ForcedAlignmentResult>{ServerError(ServerError::InvalidData, msg)};
-    }
-    if (transcript.empty()) {
-        std::string msg = "forcedAlignment requires non-empty transcript";
-        if (span)
-            span->setError(msg);
-        return Result<ForcedAlignmentResult>{ServerError(ServerError::InvalidData, msg)};
-    }
-
-    ElevenLabsCall call(apiKey, "https://api.elevenlabs.io/v1/forced-alignment");
-    if (!call.initOk()) {
-        std::string msg = "Failed to initialize curl";
-        if (span)
-            span->setError(msg);
-        return Result<ForcedAlignmentResult>{ServerError(ServerError::InternalError, msg)};
-    }
-    call.addHeader("Accept: application/json");
-
-    // Multipart body via curl's MIME API. Content-Type is set automatically by
-    // CURLOPT_MIMEPOST (it includes the boundary), so we don't add one ourselves.
-    curl_mime *mime = call.createMime();
-    curl_mimepart *filePart = curl_mime_addpart(mime);
-    curl_mime_name(filePart, "file");
-    curl_mime_filename(filePart, "audio.wav");
-    curl_mime_type(filePart, contentType.c_str());
-    curl_mime_data(filePart, reinterpret_cast<const char *>(audio.data()), audio.size());
-
-    curl_mimepart *textPart = curl_mime_addpart(mime);
-    curl_mime_name(textPart, "text");
-    curl_mime_data(textPart, transcript.data(), transcript.size());
-
-    std::string respBuf;
-    curl_easy_setopt(call.handle(), CURLOPT_WRITEFUNCTION, &appendToString);
-    curl_easy_setopt(call.handle(), CURLOPT_WRITEDATA, &respBuf);
-    curl_easy_setopt(call.handle(), CURLOPT_TIMEOUT, 90L);
-
-    long httpCode = 0;
-    const CURLcode res = call.perform(httpCode);
-
-    if (auto err = checkElevenLabsResponse<ForcedAlignmentResult>(res, httpCode, "ElevenLabs forced-alignment", respBuf,
-                                                                  span)) {
-        return *err;
-    }
-
-    nlohmann::json json;
-    try {
-        json = nlohmann::json::parse(respBuf);
-    } catch (const nlohmann::json::exception &e) {
-        std::string msg = fmt::format("ElevenLabs forced-alignment: response was not valid JSON: {}", e.what());
-        error(msg);
-        if (span)
-            span->setError(msg);
-        return Result<ForcedAlignmentResult>{ServerError(ServerError::InternalError, msg)};
-    }
-
-    ForcedAlignmentResult fa;
-
-    auto extractEntries = [](const nlohmann::json &arr, auto &outVec, auto makeEntry) {
-        for (const auto &item : arr) {
-            if (!item.is_object())
-                continue;
-            outVec.push_back(makeEntry(item));
-        }
-    };
-
-    if (json.contains("words") && json["words"].is_array()) {
-        extractEntries(json["words"], fa.words, [](const nlohmann::json &it) {
-            ForcedAlignmentWord w;
-            if (it.contains("text"))
-                w.text = it["text"].get<std::string>();
-            if (it.contains("start"))
-                w.startSeconds = it["start"].get<double>();
-            if (it.contains("end"))
-                w.endSeconds = it["end"].get<double>();
-            return w;
-        });
-    }
-    if (json.contains("characters") && json["characters"].is_array()) {
-        extractEntries(json["characters"], fa.characters, [](const nlohmann::json &it) {
-            ForcedAlignmentChar c;
-            if (it.contains("text"))
-                c.text = it["text"].get<std::string>();
-            if (it.contains("start"))
-                c.startSeconds = it["start"].get<double>();
-            if (it.contains("end"))
-                c.endSeconds = it["end"].get<double>();
-            return c;
-        });
-    }
-    if (json.contains("loss") && json["loss"].is_number()) {
-        fa.loss = json["loss"].get<double>();
-    }
-
-    if (fa.characters.empty() && fa.words.empty()) {
-        std::string msg = "ElevenLabs forced-alignment: response had no words or characters";
-        error(msg);
-        if (span)
-            span->setError(msg);
-        return Result<ForcedAlignmentResult>{ServerError(ServerError::InternalError, msg)};
-    }
-
-    info("Forced alignment complete: {} words, {} chars, loss={:.3f}", fa.words.size(), fa.characters.size(), fa.loss);
-
-    if (span) {
-        span->setAttribute("forced_alignment.words", static_cast<int64_t>(fa.words.size()));
-        span->setAttribute("forced_alignment.chars", static_cast<int64_t>(fa.characters.size()));
-        span->setAttribute("forced_alignment.loss", fa.loss);
-        span->setSuccess();
-    }
-    return fa;
 }
 
 } // namespace creatures::voice
